requirements.txt
python==3.7.6
torch==1.12.0+cu116
torchaudio==0.12.0+cu116
torchvision==0.13.0+cu116
numpy
pandas
requests
tqdm
scipy
Pillow